pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forecast-in-a-box"
version = "0.0.1"
description = "Weather Forecast in a Box"
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]
license = { text = "Apache License Version 2.0" }
requires-python=">=3.11,<3.12"
# just a minimal set of dependencies to run the servers. Individual tasks install their runtime dependencies on their own
dependencies = [
	"fastapi",
	"httpx",
	"jinja2",
	"orjson",
	"python-multipart",
	"uvicorn",
	"cloudpickle",
	"pyrsistent",
	"anemoi-inference[huggingface]>=0.4.10",
	"qubed",
	"cascade-python",
	"anemoi-cascade",
	"pydantic-settings",
]

[project.urls]
"Source code" = "https://github.com/ecmwf/forecast-in-a-box"

[project.optional-dependencies]
thermo = [
    "thermofeel>=2.1.1",
]
plots = [
    "earthkit-plots-default-styles",
    "earthkit-plots>=0.2.8",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/forecastbox"]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
ignore = ["E731"]

[tool.ruff.format]
indent-style = "tab"

[tool.mypy]
plugins = "pydantic.mypy"

[tool.uv.sources]
cascade-python = { git = "https://github.com/ecmwf/cascade", rev = "FIAB" }
qubed = { git = "ssh://git@github.com/ecmwf/qubed.git" }
earthkit-plots = { git = "ssh://git@github.com/ecmwf/earthkit-plots", rev = "experimental" }
earthkit-plots-default-styles = { git = "ssh://git@github.com/ecmwf/earthkit-plots-default-styles", rev = "experimental" }
anemoi-cascade = { git = "ssh://git@github.com/ecmwf/anemoi-cascade.git", rev = "fix/move_to_basic_objects" }

[tool.hatch.metadata]
allow-direct-references = true
